Add functionality according to issue #2.
Added support for automated attribute release policy for entity categories.

The proxy checks the entity category(ec) for SP and chooses a SP in the back end with the same ec to communicate with the underlying IdP.

Metadata for the proxy must be generated with make_proxy_metadata.py instead of make_metadata.py.